+++ src/Neo.VM/ExecutionEngine.cs
@@ -12,18 +12,17 @@
 using Neo.VM.Types;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
-using System.Numerics;
 using System.Runtime.CompilerServices;
-using Buffer = Neo.VM.Types.Buffer;
-using VMArray = Neo.VM.Types.Array;
 
 namespace Neo.VM
 {
     /// <summary>
     /// Represents the VM used to execute the script.
     /// </summary>
-    public class ExecutionEngine : IDisposable
+    public partial class ExecutionEngine : IDisposable
     {
         private VMState state = VMState.BREAK;
         private bool isJumping = false;
@@ -159,41 +158,25 @@ namespace Neo.VM
                 case OpCode.PUSHINT64:
                 case OpCode.PUSHINT128:
                 case OpCode.PUSHINT256:
-                    {
-                        Push(new BigInteger(instruction.Operand.Span));
-                        break;
-                    }
+                    ExecutePUSHINT(instruction);
+                    break;
                 case OpCode.PUSHT:
-                    {
-                        Push(StackItem.True);
-                        break;
-                    }
+                    ExecutePUSHT(instruction);
+                    break;
                 case OpCode.PUSHF:
-                    {
-                        Push(StackItem.False);
-                        break;
-                    }
+                    ExecutePUSHF(instruction);
+                    break;
                 case OpCode.PUSHA:
-                    {
-                        int position = checked(CurrentContext!.InstructionPointer + instruction.TokenI32);
-                        if (position < 0 || position > CurrentContext.Script.Length)
-                            throw new InvalidOperationException($"Bad pointer address: {position}");
-                        Push(new Pointer(CurrentContext.Script, position));
-                        break;
-                    }
+                    ExecutePUSHA(instruction);
+                    break;
                 case OpCode.PUSHNULL:
-                    {
-                        Push(StackItem.Null);
-                        break;
-                    }
+                    ExecutePUSHNULL(instruction);
+                    break;
                 case OpCode.PUSHDATA1:
                 case OpCode.PUSHDATA2:
                 case OpCode.PUSHDATA4:
-                    {
-                        Limits.AssertMaxItemSize(instruction.Operand.Length);
-                        Push(instruction.Operand);
-                        break;
-                    }
+                    ExecutePUSHDATA(instruction);
+                    break;
                 case OpCode.PUSHM1:
                 case OpCode.PUSH0:
                 case OpCode.PUSH1:
@@ -212,367 +195,162 @@ namespace Neo.VM
                 case OpCode.PUSH14:
                 case OpCode.PUSH15:
                 case OpCode.PUSH16:
-                    {
-                        Push((int)instruction.OpCode - (int)OpCode.PUSH0);
-                        break;
-                    }
+                    ExecutePUSH(instruction);
+                    break;
 
                 // Control
                 case OpCode.NOP: break;
                 case OpCode.JMP:
-                    {
-                        ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMP(instruction);
+                    break;
                 case OpCode.JMP_L:
-                    {
-                        ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMP_L(instruction);
+                    break;
                 case OpCode.JMPIF:
-                    {
-                        if (Pop().GetBoolean())
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPIF(instruction);
+                    break;
                 case OpCode.JMPIF_L:
-                    {
-                        if (Pop().GetBoolean())
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPIF_L(instruction);
+                    break;
                 case OpCode.JMPIFNOT:
-                    {
-                        if (!Pop().GetBoolean())
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPIFNOT(instruction);
+                    break;
                 case OpCode.JMPIFNOT_L:
-                    {
-                        if (!Pop().GetBoolean())
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPIFNOT_L(instruction);
+                    break;
                 case OpCode.JMPEQ:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 == x2)
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPEQ(instruction);
+                    break;
                 case OpCode.JMPEQ_L:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 == x2)
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPEQ_L(instruction);
+                    break;
                 case OpCode.JMPNE:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 != x2)
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPNE(instruction);
+                    break;
                 case OpCode.JMPNE_L:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 != x2)
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPNE_L(instruction);
+                    break;
                 case OpCode.JMPGT:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 > x2)
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPGT(instruction);
+                    break;
                 case OpCode.JMPGT_L:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 > x2)
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPGT_L(instruction);
+                    break;
                 case OpCode.JMPGE:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 >= x2)
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPGE(instruction);
+                    break;
                 case OpCode.JMPGE_L:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 >= x2)
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPGE_L(instruction);
+                    break;
                 case OpCode.JMPLT:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 < x2)
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPLT(instruction);
+                    break;
                 case OpCode.JMPLT_L:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 < x2)
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPLT_L(instruction);
+                    break;
                 case OpCode.JMPLE:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 <= x2)
-                            ExecuteJumpOffset(instruction.TokenI8);
-                        break;
-                    }
+                    ExecuteJMPLE(instruction);
+                    break;
                 case OpCode.JMPLE_L:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        if (x1 <= x2)
-                            ExecuteJumpOffset(instruction.TokenI32);
-                        break;
-                    }
+                    ExecuteJMPLE_L(instruction);
+                    break;
                 case OpCode.CALL:
-                    {
-                        ExecuteCall(checked(CurrentContext!.InstructionPointer + instruction.TokenI8));
-                        break;
-                    }
+                    ExecuteCALL(instruction);
+                    break;
                 case OpCode.CALL_L:
-                    {
-                        ExecuteCall(checked(CurrentContext!.InstructionPointer + instruction.TokenI32));
-                        break;
-                    }
+                    ExecuteCALL_L(instruction);
+                    break;
                 case OpCode.CALLA:
-                    {
-                        var x = Pop<Pointer>();
-                        if (x.Script != CurrentContext!.Script)
-                            throw new InvalidOperationException("Pointers can't be shared between scripts");
-                        ExecuteCall(x.Position);
-                        break;
-                    }
+                    ExecuteCALLA(instruction);
+                    break;
                 case OpCode.CALLT:
-                    {
-                        LoadToken(instruction.TokenU16);
-                        break;
-                    }
+                    ExecuteCALLT(instruction);
+                    break;
                 case OpCode.ABORT:
-                    {
-                        throw new Exception($"{OpCode.ABORT} is executed.");
-                    }
+                    ExecuteABORT(instruction);
+                    break;
                 case OpCode.ASSERT:
-                    {
-                        var x = Pop().GetBoolean();
-                        if (!x)
-                            throw new Exception($"{OpCode.ASSERT} is executed with false result.");
-                        break;
-                    }
+                    ExecuteASSERT(instruction);
+                    break;
                 case OpCode.THROW:
-                    {
-                        ExecuteThrow(Pop());
-                        break;
-                    }
+                    ExecuteTHROW(instruction);
+                    break;
                 case OpCode.TRY:
-                    {
-                        int catchOffset = instruction.TokenI8;
-                        int finallyOffset = instruction.TokenI8_1;
-                        ExecuteTry(catchOffset, finallyOffset);
-                        break;
-                    }
+                    ExecuteTRY(instruction);
+                    break;
                 case OpCode.TRY_L:
-                    {
-                        int catchOffset = instruction.TokenI32;
-                        int finallyOffset = instruction.TokenI32_1;
-                        ExecuteTry(catchOffset, finallyOffset);
-                        break;
-                    }
+                    ExecuteTRY_L(instruction);
+                    break;
                 case OpCode.ENDTRY:
-                    {
-                        int endOffset = instruction.TokenI8;
-                        ExecuteEndTry(endOffset);
-                        break;
-                    }
+                    ExecuteENDTRY(instruction);
+                    break;
                 case OpCode.ENDTRY_L:
-                    {
-                        int endOffset = instruction.TokenI32;
-                        ExecuteEndTry(endOffset);
-                        break;
-                    }
+                    ExecuteENDTRY_L(instruction);
+                    break;
                 case OpCode.ENDFINALLY:
-                    {
-                        if (CurrentContext!.TryStack is null)
-                            throw new InvalidOperationException($"The corresponding TRY block cannot be found.");
-                        if (!CurrentContext.TryStack.TryPop(out ExceptionHandlingContext? currentTry))
-                            throw new InvalidOperationException($"The corresponding TRY block cannot be found.");
-
-                        if (UncaughtException is null)
-                            CurrentContext.InstructionPointer = currentTry.EndPointer;
-                        else
-                            HandleException();
-
-                        isJumping = true;
-                        break;
-                    }
+                    ExecuteENDFINALLY(instruction);
+                    break;
                 case OpCode.RET:
-                    {
-                        ExecutionContext context_pop = InvocationStack.Pop();
-                        EvaluationStack stack_eval = InvocationStack.Count == 0 ? ResultStack : InvocationStack.Peek().EvaluationStack;
-                        if (context_pop.EvaluationStack != stack_eval)
-                        {
-                            if (context_pop.RVCount >= 0 && context_pop.EvaluationStack.Count != context_pop.RVCount)
-                                throw new InvalidOperationException("RVCount doesn't match with EvaluationStack");
-                            context_pop.EvaluationStack.CopyTo(stack_eval);
-                        }
-                        if (InvocationStack.Count == 0)
-                            State = VMState.HALT;
-                        ContextUnloaded(context_pop);
-                        isJumping = true;
-                        break;
-                    }
+                    ExecuteRET(instruction);
+                    break;
                 case OpCode.SYSCALL:
-                    {
-                        OnSysCall(instruction.TokenU32);
-                        break;
-                    }
+                    ExecuteSYSCALL(instruction);
+                    break;
 
                 // Stack ops
                 case OpCode.DEPTH:
-                    {
-                        Push(CurrentContext!.EvaluationStack.Count);
-                        break;
-                    }
+                    ExecuteDEPTH(instruction);
+                    break;
                 case OpCode.DROP:
-                    {
-                        Pop();
-                        break;
-                    }
+                    ExecuteDROP(instruction);
+                    break;
                 case OpCode.NIP:
-                    {
-                        CurrentContext!.EvaluationStack.Remove<StackItem>(1);
-                        break;
-                    }
+                    ExecuteNIP(instruction);
+                    break;
                 case OpCode.XDROP:
-                    {
-                        int n = (int)Pop().GetInteger();
-                        if (n < 0)
-                            throw new InvalidOperationException($"The negative value {n} is invalid for OpCode.{instruction.OpCode}.");
-                        CurrentContext!.EvaluationStack.Remove<StackItem>(n);
-                        break;
-                    }
+                    ExecuteXDROP(instruction);
+                    break;
                 case OpCode.CLEAR:
-                    {
-                        CurrentContext!.EvaluationStack.Clear();
-                        break;
-                    }
+                    ExecuteCLEAR(instruction);
+                    break;
                 case OpCode.DUP:
-                    {
-                        Push(Peek());
-                        break;
-                    }
+                    ExecuteDUP(instruction);
+                    break;
                 case OpCode.OVER:
-                    {
-                        Push(Peek(1));
-                        break;
-                    }
+                    ExecuteOVER(instruction);
+                    break;
                 case OpCode.PICK:
-                    {
-                        int n = (int)Pop().GetInteger();
-                        if (n < 0)
-                            throw new InvalidOperationException($"The negative value {n} is invalid for OpCode.{instruction.OpCode}.");
-                        Push(Peek(n));
-                        break;
-                    }
+                    ExecutePICK(instruction);
+                    break;
                 case OpCode.TUCK:
-                    {
-                        CurrentContext!.EvaluationStack.Insert(2, Peek());
-                        break;
-                    }
+                    ExecuteTUCK(instruction);
+                    break;
                 case OpCode.SWAP:
-                    {
-                        var x = CurrentContext!.EvaluationStack.Remove<StackItem>(1);
-                        Push(x);
-                        break;
-                    }
+                    ExecuteSWAP(instruction);
+                    break;
                 case OpCode.ROT:
-                    {
-                        var x = CurrentContext!.EvaluationStack.Remove<StackItem>(2);
-                        Push(x);
-                        break;
-                    }
+                    ExecuteROT(instruction);
+                    break;
                 case OpCode.ROLL:
-                    {
-                        int n = (int)Pop().GetInteger();
-                        if (n < 0)
-                            throw new InvalidOperationException($"The negative value {n} is invalid for OpCode.{instruction.OpCode}.");
-                        if (n == 0) break;
-                        var x = CurrentContext!.EvaluationStack.Remove<StackItem>(n);
-                        Push(x);
-                        break;
-                    }
+                    ExecuteROLL(instruction);
+                    break;
                 case OpCode.REVERSE3:
-                    {
-                        CurrentContext!.EvaluationStack.Reverse(3);
-                        break;
-                    }
+                    ExecuteREVERSE3(instruction);
+                    break;
                 case OpCode.REVERSE4:
-                    {
-                        CurrentContext!.EvaluationStack.Reverse(4);
-                        break;
-                    }
+                    ExecuteREVERSE4(instruction);
+                    break;
                 case OpCode.REVERSEN:
-                    {
-                        int n = (int)Pop().GetInteger();
-                        CurrentContext!.EvaluationStack.Reverse(n);
-                        break;
-                    }
+                    ExecuteREVERSEN(instruction);
+                    break;
 
                 //Slot
                 case OpCode.INITSSLOT:
-                    {
-                        if (CurrentContext!.StaticFields != null)
-                            throw new InvalidOperationException($"{instruction.OpCode} cannot be executed twice.");
-                        if (instruction.TokenU8 == 0)
-                            throw new InvalidOperationException($"The operand {instruction.TokenU8} is invalid for OpCode.{instruction.OpCode}.");
-                        CurrentContext.StaticFields = new Slot(instruction.TokenU8, ReferenceCounter);
-                        break;
-                    }
+                    ExecuteINITSSLOT(instruction);
+                    break;
                 case OpCode.INITSLOT:
-                    {
-                        if (CurrentContext!.LocalVariables != null || CurrentContext.Arguments != null)
-                            throw new InvalidOperationException($"{instruction.OpCode} cannot be executed twice.");
-                        if (instruction.TokenU16 == 0)
-                            throw new InvalidOperationException($"The operand {instruction.TokenU16} is invalid for OpCode.{instruction.OpCode}.");
-                        if (instruction.TokenU8 > 0)
-                        {
-                            CurrentContext.LocalVariables = new Slot(instruction.TokenU8, ReferenceCounter);
-                        }
-                        if (instruction.TokenU8_1 > 0)
-                        {
-                            StackItem[] items = new StackItem[instruction.TokenU8_1];
-                            for (int i = 0; i < instruction.TokenU8_1; i++)
-                            {
-                                items[i] = Pop();
-                            }
-                            CurrentContext.Arguments = new Slot(items, ReferenceCounter);
-                        }
-                        break;
-                    }
+                    ExecuteINITSLOT(instruction);
+                    break;
                 case OpCode.LDSFLD0:
                 case OpCode.LDSFLD1:
                 case OpCode.LDSFLD2:
@@ -580,15 +358,11 @@ namespace Neo.VM
                 case OpCode.LDSFLD4:
                 case OpCode.LDSFLD5:
                 case OpCode.LDSFLD6:
-                    {
-                        ExecuteLoadFromSlot(CurrentContext!.StaticFields, instruction.OpCode - OpCode.LDSFLD0);
-                        break;
-                    }
+                    ExecuteLDSFLDM(instruction);
+                    break;
                 case OpCode.LDSFLD:
-                    {
-                        ExecuteLoadFromSlot(CurrentContext!.StaticFields, instruction.TokenU8);
-                        break;
-                    }
+                    ExecuteLDSFLD(instruction);
+                    break;
                 case OpCode.STSFLD0:
                 case OpCode.STSFLD1:
                 case OpCode.STSFLD2:
@@ -596,15 +370,11 @@ namespace Neo.VM
                 case OpCode.STSFLD4:
                 case OpCode.STSFLD5:
                 case OpCode.STSFLD6:
-                    {
-                        ExecuteStoreToSlot(CurrentContext!.StaticFields, instruction.OpCode - OpCode.STSFLD0);
-                        break;
-                    }
+                    ExecuteSTSFLDM(instruction);
+                    break;
                 case OpCode.STSFLD:
-                    {
-                        ExecuteStoreToSlot(CurrentContext!.StaticFields, instruction.TokenU8);
-                        break;
-                    }
+                    ExecuteSTSFLD(instruction);
+                    break;
                 case OpCode.LDLOC0:
                 case OpCode.LDLOC1:
                 case OpCode.LDLOC2:
@@ -612,15 +382,11 @@ namespace Neo.VM
                 case OpCode.LDLOC4:
                 case OpCode.LDLOC5:
                 case OpCode.LDLOC6:
-                    {
-                        ExecuteLoadFromSlot(CurrentContext!.LocalVariables, instruction.OpCode - OpCode.LDLOC0);
-                        break;
-                    }
+                    ExecuteLDLOCM(instruction);
+                    break;
                 case OpCode.LDLOC:
-                    {
-                        ExecuteLoadFromSlot(CurrentContext!.LocalVariables, instruction.TokenU8);
-                        break;
-                    }
+                    ExecuteLDLOC(instruction);
+                    break;
                 case OpCode.STLOC0:
                 case OpCode.STLOC1:
                 case OpCode.STLOC2:
@@ -628,15 +394,11 @@ namespace Neo.VM
                 case OpCode.STLOC4:
                 case OpCode.STLOC5:
                 case OpCode.STLOC6:
-                    {
-                        ExecuteStoreToSlot(CurrentContext!.LocalVariables, instruction.OpCode - OpCode.STLOC0);
-                        break;
-                    }
+                    ExecuteSTLOCM(instruction);
+                    break;
                 case OpCode.STLOC:
-                    {
-                        ExecuteStoreToSlot(CurrentContext!.LocalVariables, instruction.TokenU8);
-                        break;
-                    }
+                    ExecuteSTLOC(instruction);
+                    break;
                 case OpCode.LDARG0:
                 case OpCode.LDARG1:
                 case OpCode.LDARG2:
@@ -644,15 +406,11 @@ namespace Neo.VM
                 case OpCode.LDARG4:
                 case OpCode.LDARG5:
                 case OpCode.LDARG6:
-                    {
-                        ExecuteLoadFromSlot(CurrentContext!.Arguments, instruction.OpCode - OpCode.LDARG0);
-                        break;
-                    }
+                    ExecuteLDARGM(instruction);
+                    break;
                 case OpCode.LDARG:
-                    {
-                        ExecuteLoadFromSlot(CurrentContext!.Arguments, instruction.TokenU8);
-                        break;
-                    }
+                    ExecuteLDARG(instruction);
+                    break;
                 case OpCode.STARG0:
                 case OpCode.STARG1:
                 case OpCode.STARG2:
@@ -660,742 +418,222 @@ namespace Neo.VM
                 case OpCode.STARG4:
                 case OpCode.STARG5:
                 case OpCode.STARG6:
-                    {
-                        ExecuteStoreToSlot(CurrentContext!.Arguments, instruction.OpCode - OpCode.STARG0);
-                        break;
-                    }
+                    ExecuteSTARGM(instruction);
+                    break;
                 case OpCode.STARG:
-                    {
-                        ExecuteStoreToSlot(CurrentContext!.Arguments, instruction.TokenU8);
-                        break;
-                    }
+                    ExecuteSTARG(instruction);
+                    break;
 
                 // Splice
                 case OpCode.NEWBUFFER:
-                    {
-                        int length = (int)Pop().GetInteger();
-                        Limits.AssertMaxItemSize(length);
-                        Push(new Buffer(length));
-                        break;
-                    }
+                    ExecuteNEWBUFFER(instruction);
+                    break;
                 case OpCode.MEMCPY:
-                    {
-                        int count = (int)Pop().GetInteger();
-                        if (count < 0)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        int si = (int)Pop().GetInteger();
-                        if (si < 0)
-                            throw new InvalidOperationException($"The value {si} is out of range.");
-                        ReadOnlySpan<byte> src = Pop().GetSpan();
-                        if (checked(si + count) > src.Length)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        int di = (int)Pop().GetInteger();
-                        if (di < 0)
-                            throw new InvalidOperationException($"The value {di} is out of range.");
-                        Buffer dst = Pop<Buffer>();
-                        if (checked(di + count) > dst.Size)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        src.Slice(si, count).CopyTo(dst.InnerBuffer.Span[di..]);
-                        break;
-                    }
+                    ExecuteMEMCPY(instruction);
+                    break;
                 case OpCode.CAT:
-                    {
-                        var x2 = Pop().GetSpan();
-                        var x1 = Pop().GetSpan();
-                        int length = x1.Length + x2.Length;
-                        Limits.AssertMaxItemSize(length);
-                        Buffer result = new(length, false);
-                        x1.CopyTo(result.InnerBuffer.Span);
-                        x2.CopyTo(result.InnerBuffer.Span[x1.Length..]);
-                        Push(result);
-                        break;
-                    }
+                    ExecuteCAT(instruction);
+                    break;
                 case OpCode.SUBSTR:
-                    {
-                        int count = (int)Pop().GetInteger();
-                        if (count < 0)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        int index = (int)Pop().GetInteger();
-                        if (index < 0)
-                            throw new InvalidOperationException($"The value {index} is out of range.");
-                        var x = Pop().GetSpan();
-                        if (index + count > x.Length)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        Buffer result = new(count, false);
-                        x.Slice(index, count).CopyTo(result.InnerBuffer.Span);
-                        Push(result);
-                        break;
-                    }
+                    ExecuteSUBSTR(instruction);
+                    break;
                 case OpCode.LEFT:
-                    {
-                        int count = (int)Pop().GetInteger();
-                        if (count < 0)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        var x = Pop().GetSpan();
-                        if (count > x.Length)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        Buffer result = new(count, false);
-                        x[..count].CopyTo(result.InnerBuffer.Span);
-                        Push(result);
-                        break;
-                    }
+                    ExecuteLEFT(instruction);
+                    break;
                 case OpCode.RIGHT:
-                    {
-                        int count = (int)Pop().GetInteger();
-                        if (count < 0)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        var x = Pop().GetSpan();
-                        if (count > x.Length)
-                            throw new InvalidOperationException($"The value {count} is out of range.");
-                        Buffer result = new(count, false);
-                        x[^count..^0].CopyTo(result.InnerBuffer.Span);
-                        Push(result);
-                        break;
-                    }
+                    ExecuteRIGHT(instruction);
+                    break;
 
                 // Bitwise logic
                 case OpCode.INVERT:
-                    {
-                        var x = Pop().GetInteger();
-                        Push(~x);
-                        break;
-                    }
+                    ExecuteINVERT(instruction);
+                    break;
                 case OpCode.AND:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 & x2);
-                        break;
-                    }
+                    ExecuteAND(instruction);
+                    break;
                 case OpCode.OR:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 | x2);
-                        break;
-                    }
+                    ExecuteOR(instruction);
+                    break;
                 case OpCode.XOR:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 ^ x2);
-                        break;
-                    }
+                    ExecuteXOR(instruction);
+                    break;
                 case OpCode.EQUAL:
-                    {
-                        StackItem x2 = Pop();
-                        StackItem x1 = Pop();
-                        Push(x1.Equals(x2, Limits));
-                        break;
-                    }
+                    ExecuteEQUAL(instruction);
+                    break;
                 case OpCode.NOTEQUAL:
-                    {
-                        StackItem x2 = Pop();
-                        StackItem x1 = Pop();
-                        Push(!x1.Equals(x2, Limits));
-                        break;
-                    }
+                    ExecuteNOTEQUAL(instruction);
+                    break;
 
                 // Numeric
                 case OpCode.SIGN:
-                    {
-                        var x = Pop().GetInteger();
-                        Push(x.Sign);
-                        break;
-                    }
+                    ExecuteSIGN(instruction);
+                    break;
                 case OpCode.ABS:
-                    {
-                        var x = Pop().GetInteger();
-                        Push(BigInteger.Abs(x));
-                        break;
-                    }
+                    ExecuteABS(instruction);
+                    break;
                 case OpCode.NEGATE:
-                    {
-                        var x = Pop().GetInteger();
-                        Push(-x);
-                        break;
-                    }
+                    ExecuteNEGATE(instruction);
+                    break;
                 case OpCode.INC:
-                    {
-                        var x = Pop().GetInteger();
-                        Push(x + 1);
-                        break;
-                    }
+                    ExecuteINC(instruction);
+                    break;
                 case OpCode.DEC:
-                    {
-                        var x = Pop().GetInteger();
-                        Push(x - 1);
-                        break;
-                    }
+                    ExecuteDEC(instruction);
+                    break;
                 case OpCode.ADD:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 + x2);
-                        break;
-                    }
+                    ExecuteADD(instruction);
+                    break;
                 case OpCode.SUB:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 - x2);
-                        break;
-                    }
+                    ExecuteSUB(instruction);
+                    break;
                 case OpCode.MUL:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 * x2);
-                        break;
-                    }
+                    ExecuteMUL(instruction);
+                    break;
                 case OpCode.DIV:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 / x2);
-                        break;
-                    }
+                    ExecuteDIV(instruction);
+                    break;
                 case OpCode.MOD:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 % x2);
-                        break;
-                    }
+                    ExecuteMOD(instruction);
+                    break;
                 case OpCode.POW:
-                    {
-                        var exponent = (int)Pop().GetInteger();
-                        Limits.AssertShift(exponent);
-                        var value = Pop().GetInteger();
-                        Push(BigInteger.Pow(value, exponent));
-                        break;
-                    }
+                    ExecutePOW(instruction);
+                    break;
                 case OpCode.SQRT:
-                    {
-                        Push(Pop().GetInteger().Sqrt());
-                        break;
-                    }
+                    ExecuteSQRT(instruction);
+                    break;
                 case OpCode.MODMUL:
-                    {
-                        var modulus = Pop().GetInteger();
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 * x2 % modulus);
-                        break;
-                    }
+                    ExecuteMODMUL(instruction);
+                    break;
                 case OpCode.MODPOW:
-                    {
-                        var modulus = Pop().GetInteger();
-                        var exponent = Pop().GetInteger();
-                        var value = Pop().GetInteger();
-                        var result = exponent == -1
-                            ? value.ModInverse(modulus)
-                            : BigInteger.ModPow(value, exponent, modulus);
-                        Push(result);
-                        break;
-                    }
+                    ExecuteMODPOW(instruction);
+                    break;
                 case OpCode.SHL:
-                    {
-                        int shift = (int)Pop().GetInteger();
-                        Limits.AssertShift(shift);
-                        if (shift == 0) break;
-                        var x = Pop().GetInteger();
-                        Push(x << shift);
-                        break;
-                    }
+                    ExecuteSHL(instruction);
+                    break;
                 case OpCode.SHR:
-                    {
-                        int shift = (int)Pop().GetInteger();
-                        Limits.AssertShift(shift);
-                        if (shift == 0) break;
-                        var x = Pop().GetInteger();
-                        Push(x >> shift);
-                        break;
-                    }
+                    ExecuteSHR(instruction);
+                    break;
                 case OpCode.NOT:
-                    {
-                        var x = Pop().GetBoolean();
-                        Push(!x);
-                        break;
-                    }
+                    ExecuteNOT(instruction);
+                    break;
                 case OpCode.BOOLAND:
-                    {
-                        var x2 = Pop().GetBoolean();
-                        var x1 = Pop().GetBoolean();
-                        Push(x1 && x2);
-                        break;
-                    }
+                    ExecuteBOOLAND(instruction);
+                    break;
                 case OpCode.BOOLOR:
-                    {
-                        var x2 = Pop().GetBoolean();
-                        var x1 = Pop().GetBoolean();
-                        Push(x1 || x2);
-                        break;
-                    }
+                    ExecuteBOOLOR(instruction);
+                    break;
                 case OpCode.NZ:
-                    {
-                        var x = Pop().GetInteger();
-                        Push(!x.IsZero);
-                        break;
-                    }
+                    ExecuteNZ(instruction);
+                    break;
                 case OpCode.NUMEQUAL:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 == x2);
-                        break;
-                    }
+                    ExecuteNUMEQUAL(instruction);
+                    break;
                 case OpCode.NUMNOTEQUAL:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(x1 != x2);
-                        break;
-                    }
+                    ExecuteNUMNOTEQUAL(instruction);
+                    break;
                 case OpCode.LT:
-                    {
-                        var x2 = Pop();
-                        var x1 = Pop();
-                        if (x1.IsNull || x2.IsNull)
-                            Push(false);
-                        else
-                            Push(x1.GetInteger() < x2.GetInteger());
-                        break;
-                    }
+                    ExecuteLT(instruction);
+                    break;
                 case OpCode.LE:
-                    {
-                        var x2 = Pop();
-                        var x1 = Pop();
-                        if (x1.IsNull || x2.IsNull)
-                            Push(false);
-                        else
-                            Push(x1.GetInteger() <= x2.GetInteger());
-                        break;
-                    }
+                    ExecuteLE(instruction);
+                    break;
                 case OpCode.GT:
-                    {
-                        var x2 = Pop();
-                        var x1 = Pop();
-                        if (x1.IsNull || x2.IsNull)
-                            Push(false);
-                        else
-                            Push(x1.GetInteger() > x2.GetInteger());
-                        break;
-                    }
+                    ExecuteGT(instruction);
+                    break;
                 case OpCode.GE:
-                    {
-                        var x2 = Pop();
-                        var x1 = Pop();
-                        if (x1.IsNull || x2.IsNull)
-                            Push(false);
-                        else
-                            Push(x1.GetInteger() >= x2.GetInteger());
-                        break;
-                    }
+                    ExecuteGE(instruction);
+                    break;
                 case OpCode.MIN:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(BigInteger.Min(x1, x2));
-                        break;
-                    }
+                    ExecuteMIN(instruction);
+                    break;
                 case OpCode.MAX:
-                    {
-                        var x2 = Pop().GetInteger();
-                        var x1 = Pop().GetInteger();
-                        Push(BigInteger.Max(x1, x2));
-                        break;
-                    }
+                    ExecuteMAX(instruction);
+                    break;
                 case OpCode.WITHIN:
-                    {
-                        BigInteger b = Pop().GetInteger();
-                        BigInteger a = Pop().GetInteger();
-                        var x = Pop().GetInteger();
-                        Push(a <= x && x < b);
-                        break;
-                    }
+                    ExecuteWITHIN(instruction);
+                    break;
 
                 // Compound-type
                 case OpCode.PACKMAP:
-                    {
-                        int size = (int)Pop().GetInteger();
-                        if (size < 0 || size * 2 > CurrentContext!.EvaluationStack.Count)
-                            throw new InvalidOperationException($"The value {size} is out of range.");
-                        Map map = new(ReferenceCounter);
-                        for (int i = 0; i < size; i++)
-                        {
-                            PrimitiveType key = Pop<PrimitiveType>();
-                            StackItem value = Pop();
-                            map[key] = value;
-                        }
-                        Push(map);
-                        break;
-                    }
+                    ExecutePACKMAP(instruction);
+                    break;
                 case OpCode.PACKSTRUCT:
-                    {
-                        int size = (int)Pop().GetInteger();
-                        if (size < 0 || size > CurrentContext!.EvaluationStack.Count)
-                            throw new InvalidOperationException($"The value {size} is out of range.");
-                        Struct @struct = new(ReferenceCounter);
-                        for (int i = 0; i < size; i++)
-                        {
-                            StackItem item = Pop();
-                            @struct.Add(item);
-                        }
-                        Push(@struct);
-                        break;
-                    }
+                    ExecutePACKSTRUCT(instruction);
+                    break;
                 case OpCode.PACK:
-                    {
-                        int size = (int)Pop().GetInteger();
-                        if (size < 0 || size > CurrentContext!.EvaluationStack.Count)
-                            throw new InvalidOperationException($"The value {size} is out of range.");
-                        VMArray array = new(ReferenceCounter);
-                        for (int i = 0; i < size; i++)
-                        {
-                            StackItem item = Pop();
-                            array.Add(item);
-                        }
-                        Push(array);
-                        break;
-                    }
+                    ExecutePACK(instruction);
+                    break;
                 case OpCode.UNPACK:
-                    {
-                        CompoundType compound = Pop<CompoundType>();
-                        switch (compound)
-                        {
-                            case Map map:
-                                foreach (var (key, value) in map.Reverse())
-                                {
-                                    Push(value);
-                                    Push(key);
-                                }
-                                break;
-                            case VMArray array:
-                                for (int i = array.Count - 1; i >= 0; i--)
-                                {
-                                    Push(array[i]);
-                                }
-                                break;
-                            default:
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {compound.Type}");
-                        }
-                        Push(compound.Count);
-                        break;
-                    }
+                    ExecuteUNPACK(instruction);
+                    break;
                 case OpCode.NEWARRAY0:
-                    {
-                        Push(new VMArray(ReferenceCounter));
-                        break;
-                    }
+                    ExecuteNEWARRAY0(instruction);
+                    break;
                 case OpCode.NEWARRAY:
                 case OpCode.NEWARRAY_T:
-                    {
-                        int n = (int)Pop().GetInteger();
-                        if (n < 0 || n > Limits.MaxStackSize)
-                            throw new InvalidOperationException($"MaxStackSize exceed: {n}");
-                        StackItem item;
-                        if (instruction.OpCode == OpCode.NEWARRAY_T)
-                        {
-                            StackItemType type = (StackItemType)instruction.TokenU8;
-                            if (!Enum.IsDefined(typeof(StackItemType), type))
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {instruction.TokenU8}");
-                            item = instruction.TokenU8 switch
-                            {
-                                (byte)StackItemType.Boolean => StackItem.False,
-                                (byte)StackItemType.Integer => Integer.Zero,
-                                (byte)StackItemType.ByteString => ByteString.Empty,
-                                _ => StackItem.Null
-                            };
-                        }
-                        else
-                        {
-                            item = StackItem.Null;
-                        }
-                        Push(new VMArray(ReferenceCounter, Enumerable.Repeat(item, n)));
-                        break;
-                    }
+                    ExecuteNEWARRAY_T(instruction);
+                    break;
                 case OpCode.NEWSTRUCT0:
-                    {
-                        Push(new Struct(ReferenceCounter));
-                        break;
-                    }
+                    ExecuteNEWSTRUCT0(instruction);
+                    break;
                 case OpCode.NEWSTRUCT:
-                    {
-                        int n = (int)Pop().GetInteger();
-                        if (n < 0 || n > Limits.MaxStackSize)
-                            throw new InvalidOperationException($"MaxStackSize exceed: {n}");
-                        Struct result = new(ReferenceCounter);
-                        for (var i = 0; i < n; i++)
-                            result.Add(StackItem.Null);
-                        Push(result);
-                        break;
-                    }
+                    ExecuteNEWSTRUCT(instruction);
+                    break;
                 case OpCode.NEWMAP:
-                    {
-                        Push(new Map(ReferenceCounter));
-                        break;
-                    }
+                    ExecuteNEWMAP(instruction);
+                    break;
                 case OpCode.SIZE:
-                    {
-                        var x = Pop();
-                        switch (x)
-                        {
-                            case CompoundType compound:
-                                Push(compound.Count);
-                                break;
-                            case PrimitiveType primitive:
-                                Push(primitive.Size);
-                                break;
-                            case Buffer buffer:
-                                Push(buffer.Size);
-                                break;
-                            default:
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {x.Type}");
-                        }
-                        break;
-                    }
+                    ExecuteSIZE(instruction);
+                    break;
                 case OpCode.HASKEY:
-                    {
-                        PrimitiveType key = Pop<PrimitiveType>();
-                        var x = Pop();
-                        switch (x)
-                        {
-                            case VMArray array:
-                                {
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0)
-                                        throw new InvalidOperationException($"The negative value {index} is invalid for OpCode.{instruction.OpCode}.");
-                                    Push(index < array.Count);
-                                    break;
-                                }
-                            case Map map:
-                                {
-                                    Push(map.ContainsKey(key));
-                                    break;
-                                }
-                            case Buffer buffer:
-                                {
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0)
-                                        throw new InvalidOperationException($"The negative value {index} is invalid for OpCode.{instruction.OpCode}.");
-                                    Push(index < buffer.Size);
-                                    break;
-                                }
-                            case ByteString array:
-                                {
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0)
-                                        throw new InvalidOperationException($"The negative value {index} is invalid for OpCode.{instruction.OpCode}.");
-                                    Push(index < array.Size);
-                                    break;
-                                }
-                            default:
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {x.Type}");
-                        }
-                        break;
-                    }
+                    ExecuteHASKEY(instruction);
+                    break;
                 case OpCode.KEYS:
-                    {
-                        Map map = Pop<Map>();
-                        Push(new VMArray(ReferenceCounter, map.Keys));
-                        break;
-                    }
+                    ExecuteKEYS(instruction);
+                    break;
                 case OpCode.VALUES:
-                    {
-                        var x = Pop();
-                        IEnumerable<StackItem> values = x switch
-                        {
-                            VMArray array => array,
-                            Map map => map.Values,
-                            _ => throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {x.Type}"),
-                        };
-                        VMArray newArray = new(ReferenceCounter);
-                        foreach (StackItem item in values)
-                            if (item is Struct s)
-                                newArray.Add(s.Clone(Limits));
-                            else
-                                newArray.Add(item);
-                        Push(newArray);
-                        break;
-                    }
+                    ExecuteVALUES(instruction);
+                    break;
                 case OpCode.PICKITEM:
-                    {
-                        PrimitiveType key = Pop<PrimitiveType>();
-                        var x = Pop();
-                        switch (x)
-                        {
-                            case VMArray array:
-                                {
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0 || index >= array.Count)
-                                        throw new CatchableException($"The value {index} is out of range.");
-                                    Push(array[index]);
-                                    break;
-                                }
-                            case Map map:
-                                {
-                                    if (!map.TryGetValue(key, out StackItem? value))
-                                        throw new CatchableException($"Key not found in {nameof(Map)}");
-                                    Push(value);
-                                    break;
-                                }
-                            case PrimitiveType primitive:
-                                {
-                                    ReadOnlySpan<byte> byteArray = primitive.GetSpan();
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0 || index >= byteArray.Length)
-                                        throw new CatchableException($"The value {index} is out of range.");
-                                    Push((BigInteger)byteArray[index]);
-                                    break;
-                                }
-                            case Buffer buffer:
-                                {
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0 || index >= buffer.Size)
-                                        throw new CatchableException($"The value {index} is out of range.");
-                                    Push((BigInteger)buffer.InnerBuffer.Span[index]);
-                                    break;
-                                }
-                            default:
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {x.Type}");
-                        }
-                        break;
-                    }
+                    ExecutePICKITEM(instruction);
+                    break;
                 case OpCode.APPEND:
-                    {
-                        StackItem newItem = Pop();
-                        VMArray array = Pop<VMArray>();
-                        if (newItem is Struct s) newItem = s.Clone(Limits);
-                        array.Add(newItem);
-                        break;
-                    }
+                    ExecuteAPPEND(instruction);
+                    break;
                 case OpCode.SETITEM:
-                    {
-                        StackItem value = Pop();
-                        if (value is Struct s) value = s.Clone(Limits);
-                        PrimitiveType key = Pop<PrimitiveType>();
-                        var x = Pop();
-                        switch (x)
-                        {
-                            case VMArray array:
-                                {
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0 || index >= array.Count)
-                                        throw new CatchableException($"The value {index} is out of range.");
-                                    array[index] = value;
-                                    break;
-                                }
-                            case Map map:
-                                {
-                                    map[key] = value;
-                                    break;
-                                }
-                            case Buffer buffer:
-                                {
-                                    int index = (int)key.GetInteger();
-                                    if (index < 0 || index >= buffer.Size)
-                                        throw new CatchableException($"The value {index} is out of range.");
-                                    if (value is not PrimitiveType p)
-                                        throw new InvalidOperationException($"Value must be a primitive type in {instruction.OpCode}");
-                                    int b = (int)p.GetInteger();
-                                    if (b < sbyte.MinValue || b > byte.MaxValue)
-                                        throw new InvalidOperationException($"Overflow in {instruction.OpCode}, {b} is not a byte type.");
-                                    buffer.InnerBuffer.Span[index] = (byte)b;
-                                    break;
-                                }
-                            default:
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {x.Type}");
-                        }
-                        break;
-                    }
+                    ExecuteSETITEM(instruction);
+                    break;
                 case OpCode.REVERSEITEMS:
-                    {
-                        var x = Pop();
-                        switch (x)
-                        {
-                            case VMArray array:
-                                array.Reverse();
-                                break;
-                            case Buffer buffer:
-                                buffer.InnerBuffer.Span.Reverse();
-                                break;
-                            default:
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {x.Type}");
-                        }
-                        break;
-                    }
+                    ExecuteREVERSEITEMS(instruction);
+                    break;
                 case OpCode.REMOVE:
-                    {
-                        PrimitiveType key = Pop<PrimitiveType>();
-                        var x = Pop();
-                        switch (x)
-                        {
-                            case VMArray array:
-                                int index = (int)key.GetInteger();
-                                if (index < 0 || index >= array.Count)
-                                    throw new InvalidOperationException($"The value {index} is out of range.");
-                                array.RemoveAt(index);
-                                break;
-                            case Map map:
-                                map.Remove(key);
-                                break;
-                            default:
-                                throw new InvalidOperationException($"Invalid type for {instruction.OpCode}: {x.Type}");
-                        }
-                        break;
-                    }
+                    ExecuteREMOVE(instruction);
+                    break;
                 case OpCode.CLEARITEMS:
-                    {
-                        CompoundType x = Pop<CompoundType>();
-                        x.Clear();
-                        break;
-                    }
+                    ExecuteCLEARITEMS(instruction);
+                    break;
                 case OpCode.POPITEM:
-                    {
-                        VMArray x = Pop<VMArray>();
-                        int index = x.Count - 1;
-                        Push(x[index]);
-                        x.RemoveAt(index);
-                        break;
-                    }
+                    ExecutePOPITEM(instruction);
+                    break;
 
                 //Types
                 case OpCode.ISNULL:
-                    {
-                        var x = Pop();
-                        Push(x.IsNull);
-                        break;
-                    }
+                    ExecuteISNULL(instruction);
+                    break;
                 case OpCode.ISTYPE:
-                    {
-                        var x = Pop();
-                        StackItemType type = (StackItemType)instruction.TokenU8;
-                        if (type == StackItemType.Any || !Enum.IsDefined(typeof(StackItemType), type))
-                            throw new InvalidOperationException($"Invalid type: {type}");
-                        Push(x.Type == type);
-                        break;
-                    }
+                    ExecuteISTYPE(instruction);
+                    break;
                 case OpCode.CONVERT:
-                    {
-                        var x = Pop();
-                        Push(x.ConvertTo((StackItemType)instruction.TokenU8));
-                        break;
-                    }
+                    ExecuteCONVERT(instruction);
+                    break;
                 case OpCode.ABORTMSG:
-                    {
-                        var msg = Pop().GetString();
-                        throw new Exception($"{OpCode.ABORTMSG} is executed. Reason: {msg}");
-                    }
+                    ExecuteABORTMSG(instruction);
+                    break;
                 case OpCode.ASSERTMSG:
-                    {
-                        var msg = Pop().GetString();
-                        var x = Pop().GetBoolean();
-                        if (!x)
-                            throw new Exception($"{OpCode.ASSERTMSG} is executed with false result. Reason: {msg}");
-                        break;
-                    }
-                default: throw new InvalidOperationException($"Opcode {instruction.OpCode} is undefined.");
+                    ExecuteASSERTMSG(instruction);
+                    break;
+                default:
+                    throw new InvalidOperationException($"Opcode {instruction.OpCode} is undefined.");
             }
         }
